"check_style_score -> 0 of 5 passed. style score: 0/1 "

+++ cspp1 assignments/m2/Assignment-3/Solution.java
@@ -8,7 +8,7 @@ public class Solution {
 		Scanner s=new Scanner(System.in);      
         int base = s.nextInt();
         int exponent = s.nextInt();
-        int result=(int)power(base,exponent);
+        double result=(int)power(base,exponent);
         System.out.println(result);
 	}
 	/*
